Extract restful connections_controller
