+++ app/assets/javascripts/bookingsync_portal/admin/rentals.js.coffee
@@ -1,82 +1,4 @@
 $ ->
-  $(".bookingsync-rental").draggable
-    revert: "invalid"
-    zIndex: 50
-    revertDuration: 100
-    cursor: "move"
-    containment: '.rentals-container'
-    appendTo: '.rentals-container'
-    helper: "clone"
-    scroll: false
-
-    start: (e, ui) ->
-      $(ui.helper).addClass "ui-draggable-helper"
-
-  $(".not-connected-remote-rental").droppable
-    accept: ".bookingsync-rental"
-    activeClass: "dropzone-active"
-    hoverClass: "dropzone-hover"
-    greedy: true
-    tolerance: "pointer"
-    drop: (event, ui) ->
-      remoteRental = $(@)
-
-      rentalId = parseInt($(ui.draggable).attr("id").split("_").pop())
-      remoteRentalId = parseInt(remoteRental.attr("id").split("_").pop())
-      remoteRentalUid = parseInt(remoteRental.data("uid"))
-
-      remoteRental.replaceWith HandlebarsTemplates["rentals/connected_rental"]
-        rentalName: $(ui.draggable).children('.panel-heading').text()
-        rentalDescription: $(ui.draggable).children('.panel-body').html()
-        rentalId: rentalId
-        listingId: "Listing #" + remoteRentalUid
-      $(ui.draggable).remove()
-
-      connect_url = "/en/admin/rentals/" + rentalId + "/connect" +
-        "?remote_rental_id=" + remoteRentalId
-
-      $.ajax
-        url: connect_url
-        type: "PUT"
-        dataType: 'json'
-        beforeSend: ->
-          $(@).addClass('loading')
-        success: ->
-          $(@).removeClass('loading')
-
-  $(".new-remote-rental").droppable
-    accept: ".bookingsync-rental"
-    activeClass: "dropzone-active"
-    hoverClass: "dropzone-hover"
-    greedy: true
-    tolerance: "pointer"
-    drop: (event, ui) ->
-      newRemoteRental = $(@)
-
-      rentalId = parseInt($(ui.draggable).attr("id").split("_")[1])
-      remoteAccountId = parseInt($(@).data("remote-account-id"))
-
-      newRental = HandlebarsTemplates["rentals/connected_rental"]
-        rentalName: $(ui.draggable).children('.panel-heading').text()
-        rentalDescription: $(ui.draggable).children('.panel-body').html()
-        rentalId: rentalId
-
-      $(newRental).insertBefore(newRemoteRental).addClass('pending')
-
-      $(ui.draggable).remove()
-
-      new_connect_url = "/en/admin/rentals/" + rentalId + "/connect_to_new" +
-        "?remote_account_id=" + remoteAccountId
-
-      $.ajax
-        url: new_connect_url
-        type: "PUT"
-        dataType: 'json'
-        beforeSend: ->
-          $(@).addClass('loading')
-        success: ->
-          $(@).removeClass('loading')
-
   for rentalsList, index in $(".rentals-list")
     inputId = "rentals-list-filter-#{index}"
     new ListFilter(
