+++ app/assets/javascripts/bookingsync_portal/admin/templates/rentals/connected_rental.hbs
@@ -1,8 +1,8 @@
 <div class="panel panel-connected" id="rental_{{rentalId}}"}}>
   <div class="panel-heading" title="{{listingId}}">{{rentalName}}
-    <a href="/en/admin/rentals/{{rentalId}}/disconnect"
+    <a href="{{disconnect_url}}"
       class="btn btn-default btn-xs remove-connection"
-      data-method="put">Remove Connection</a>
+      data-method="delete">Remove Connection</a>
   </div>
   <div class="panel-body"><i class="fa fa-spinner fa-spin"></i> Synchronizing...</div>
 </div>
