+++ app/controllers/bookingsync_portal/admin/rentals_controller.rb
@@ -16,36 +16,6 @@ module BookingsyncPortal
         rental
       end
 
-      def connect_to_new
-        unless BookingsyncPortal.create_remote_rental_from_app
-          raise ActionController::RoutingError.new('Not Found')
-        end
-
-        remote_account = current_account.remote_accounts.find(params[:remote_account_id])
-        remote_rental = ::RemoteRental.create!(remote_account_id: remote_account.id)
-        connection = rental.create_connection(remote_rental: remote_rental)
-
-        BookingsyncPortal.remote_rental_created(remote_rental)
-        redirect_or_js_response
-      end
-
-      def connect
-        remote_rental = current_account.remote_rentals.find(params[:remote_rental_id])
-        connection = rental.create_connection(remote_rental: remote_rental)
-
-        BookingsyncPortal.connection_created(connection)
-        redirect_or_js_response
-      end
-
-      def disconnect
-        connection = rental.connection
-        rental.remote_rental.update_attribute(:synchronized_at, nil)
-        connection.destroy
-
-        BookingsyncPortal.connection_destroyed(connection)
-        redirect_or_js_response
-      end
-
       private
 
       def synchronize_rentals
@@ -60,13 +30,6 @@ module BookingsyncPortal
 
       def rental
         @rental ||= current_account.rentals.visible.find(params[:id])
-      end
-
-      def redirect_or_js_response
-        respond_to do |wants|
-          wants.html { redirect_to admin_rentals_path }
-          wants.json { head :ok }
-        end
       end
     end
   end
