+++ app/views/bookingsync_portal/admin/rentals/_rental.html.erb
@@ -1,6 +1,5 @@
 <div class="panel panel-bookingsync bookingsync-rental"
-  id="<%= dom_id rental %>"
-  data-connect-url="<%= connect_admin_rental_url(rental) %>">
+  id="<%= dom_id rental %>">
   <div class="panel-body-grid">
     <div class="panel-body-grid-photo">
       <%- if rental.ordered_photos.first -%>
