+++ config/routes.rb
@@ -1,10 +1,7 @@
 BookingsyncPortal::Engine.routes.draw do
   namespace :admin do
-    resources :rentals, only: [:index, :show] do
-      put :disconnect, on: :member
-      put :connect, on: :member
-      put :connect_to_new, on: :member
-    end
+    resources :rentals, only: [:index, :show]
+    resources :connections, only: [:create, :destroy]
     resources :remote_accounts, only: [:new, :create]
     get 'help', to: 'help#index'
     root to: 'rentals#index'
