+++ spec/dummy/app/models/account.rb
@@ -1,5 +1,5 @@
 class Account < BookingsyncPortal::Account
-  synced local_attributes: [:email]
+  synced local_attributes: [:email], strategy: :updated_since
 
   def api
     @api ||= BookingSync::Engine::APIClient.new(token.token, logger: Rails.logger, account: self)
