+++ spec/dummy/app/models/photo.rb
@@ -1,4 +1,4 @@
 class Photo < BookingsyncPortal::Photo
   synced local_attributes: [:position], delegate_attributes: [:large_url, :thumb_url],
-    remove: true
+    remove: true, strategy: :updated_since
 end
