+++ spec/dummy/app/models/rate.rb
@@ -1,3 +1,3 @@
 class Rate < BookingsyncPortal::Rate
-  synced remove: true
+  synced remove: true, strategy: :updated_since
 end
